requirements.txt
websockets==12.0
psutil==5.9.5
requests==2.31.0
Pillow==10.1.0